moves banner
