+++ apps/hyperdrive-trading/src/ui/trade/TradeBody/TradeBody.tsx
@@ -2,6 +2,7 @@ import { ReactElement } from "react";
 import { Hyperdrive } from "src/appconfig/types";
 import { divideBigInt } from "src/base/divideBigInt";
 import { parseUnits } from "src/base/parseUnits";
+import CustomBanner from "src/ui/base/components/CustomBanner";
 import { formatBalance } from "src/ui/base/formatting/formatBalance";
 import { TransactionTable } from "src/ui/hyperdrive/TransactionTable/TransactionsTable";
 import { useTransactionData } from "src/ui/hyperdrive/TransactionTable/useTransactionData";
@@ -11,7 +12,6 @@ import { useCurrentLongPrice } from "src/ui/hyperdrive/longs/hooks/useCurrentLon
 import { AddLiquidityModalButton } from "src/ui/hyperdrive/lp/AddLiquidityModalButton/AddLiquidityModalButton";
 import { OpenShortModalButton } from "src/ui/hyperdrive/shorts/OpenShortModalButton/OpenShortModalButton";
 import { AllMarketsBreadcrumb } from "src/ui/markets/MarketSelect/AllMarketsBreadcrumb";
-import CustomBanner from "src/ui/markets/MarketStats/CustomBanner";
 import { MarketStats } from "src/ui/markets/MarketStats/MarketStats";
 import { FAQ } from "src/ui/onboarding/FAQ/FAQ";
 import { PositionsSection } from "src/ui/portfolio/PositionsSection/PositionsSection";
